module6-task2: * move table name enum
