+++ src/common/enums/db/index.ts
@@ -1,0 +1,1 @@
+export * from './table-name.enum';
