+++ src/common/enums/db/table-name.enum.ts
@@ -1,0 +1,10 @@
+enum TableName {
+  CATEGORIES = `categories`,
+  OFFER_TYPES = `offer_types`,
+  USERS = `users`,
+  COMMENTS = `comments`,
+  OFFERS = `offers`,
+  OFFERS_CATEGORIES = `offers_categories`,
+}
+
+export { TableName };
